Adding more unit tests
